M2/milestone/requirements.txt
numpy
pandas
PyYAML
requests
SQLAlchemy
psycopg2-binary
matplotlib
scikit-learn
flask
pytest
python-dotenv
pytz
seaborn
scipy
docker
kafka-python
